Use LibreOffice from command line to write out invoice PDFs.
Depends on a recent version of LibreOffice (not sure what version exactly, I'm using 5.2.0.4). For Windows version 5 is hard-coded as the default for now.
Hacking this together with a Process.waitFor() call to keep the tests passing - should really be using some kind of fake/mock process in the test but will come back to this as I don't have much free time right now and want to get a working thing up and running.